requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.6.0
beautifulsoup4==4.13.4
bs4==0.0.2
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
google==3.0.0
google-api-core==2.25.1
google-api-python-client==2.177.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
markdown-it-py==3.0.0
mcp==1.10.1
mdurl==0.1.2
oauthlib==3.3.1
openapi-pydantic==0.5.1
outcome==1.3.0.post0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
PySocks==1.7.1
python-dotenv==1.1.1
python-multipart==0.0.20
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
selenium==4.34.2
shellingham==1.5.4
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sse-starlette==2.3.6
starlette==0.47.1
trio==0.30.0
trio-websocket==0.12.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
websocket-client==1.8.0
wsproto==1.2.0
